+++ source/blender/src/header_script.c
@@ -72,7 +72,7 @@
 /* action executed after clicking in Scripts menu */
 static void do_scripts_submenus(void *int_arg, int event)
 {
-	int menutype = (int)int_arg;
+	int menutype = (long)int_arg;
 
 	BPY_menu_do_python (menutype, event);
 
@@ -84,7 +84,7 @@ static uiBlock *script_scripts_submenus(void *int_menutype)
 	uiBlock *block;
 	short yco = 20, menuwidth = 120;
 	BPyMenu *pym;
-	int i = 0, menutype = (int)int_menutype;
+	int i = 0, menutype = (long)int_menutype;
 
 	if ((menutype < 0) || (menutype > PYMENU_SCRIPTS_MENU_TOTAL))
 		return NULL;
@@ -135,7 +135,7 @@ static uiBlock *script_scriptsmenu(void *arg_unused)
 	uiBlockSetButmFunc(block, do_script_scriptsmenu, NULL);
 
 	for (i = 0; i < PYMENU_SCRIPTS_MENU_TOTAL; i++) {
-		uiDefIconTextBlockBut(block, script_scripts_submenus, (void *)i, ICON_RIGHTARROW_THIN, BPyMenu_group_itoa(i), 0, yco-=20, menuwidth, 19, "");
+		uiDefIconTextBlockBut(block, script_scripts_submenus, (void *)(long)i, ICON_RIGHTARROW_THIN, BPyMenu_group_itoa(i), 0, yco-=20, menuwidth, 19, "");
 	}
 
 	uiDefBut(block, SEPR, 0, "", 0, yco-=6, menuwidth, 6, NULL, 0.0, 0.0, 0, 0, "");
